Move add, delete, and handle events logic to separate file
